rename: api-keys -> index
